+++ core/thread.py
@@ -101,7 +101,7 @@ class Thread:
             }
         })
 
-        if not isinstance(log_data, str) or log_data is not None:
+        if log_data is not None and isinstance(log_data, dict):
             if self.bot.selfhosted:
                 log_url = f"{self.bot.config.log_url.strip('/')}/" \
                           f"logs/{log_data['key']}"
@@ -134,11 +134,15 @@ class Thread:
 
         # Thread closed message 
 
-        em = discord.Embed(title='Thread Closed', color=discord.Color.red())
+        em = discord.Embed(title='Thread Closed',
+                           color=discord.Color.red(),
+                           timestamp=datetime.utcnow())
 
         if not message:
             message = f'{closer.mention} has closed this Modmail thread.'
         em.description = message
+        em.set_footer(text='Replying will create a new thread',
+                      icon_url=self.bot.guild.icon_url)
 
         if not silent and self.recipient is not None:
             tasks.append(self.recipient.send(embed=em))
@@ -162,13 +166,11 @@ class Thread:
         async for msg in channel.history():
             if not msg.embeds:
                 continue
-            em = msg.embeds[0]
-            if em and em.author and em.author.url:
-                if str(message_id) == str(em.author.url).split('/')[-1]:
-                    if ' - (Edited)' not in em.footer.text:
-                        em.set_footer(text=em.footer.text + ' - (Edited)')
-                    em.description = message
-                    await msg.edit(embed=em)
+            embed = msg.embeds[0]
+            if embed and embed.author and embed.author.url:
+                if str(message_id) == str(embed.author.url).split('/')[-1]:
+                    embed.description = message
+                    await msg.edit(embed=embed)
                     break
 
     def edit_message(self, message_id, message):
@@ -244,18 +246,25 @@ class Thread:
 
         author = message.author
 
-        em = discord.Embed(description=message.content,
-                           timestamp=message.created_at)
+        em = discord.Embed(
+            description=message.content,
+            # timestamp=message.created_at
+        )
 
         system_avatar_url = 'https://discordapp.com/assets/' \
                             'f78426a064bc9dd24847519259bc42af.png'
 
         # store message id in hidden url
-        avatar_url = author.avatar_url if not note else system_avatar_url
-
-        em.set_author(name=str(author) if not note else 'Note',
-                      icon_url=avatar_url,
-                      url=message.jump_url)
+        if not note:
+            em.set_author(name=author.name,
+                          icon_url=author.avatar_url,
+                          url=message.jump_url)
+        else:
+            em.set_author(
+                name=f'Note ({author.name})',
+                icon_url=system_avatar_url,
+                url=message.jump_url
+            )
 
         image_types = ['.png', '.jpg', '.gif', '.jpeg', '.webp']
 
@@ -310,14 +319,11 @@ class Thread:
             file_upload_count += 1
 
         if from_mod:
-            em.color = discord.Color.green()
-            em.set_footer(text=f'Moderator')
+            em.color = self.bot.mod_color
         elif note:
             em.color = discord.Color.blurple()
-            em.set_footer(text=f'System ({author.name})')
         else:
-            em.color = discord.Color.gold()
-            em.set_footer(text=f'User')
+            em.color = self.bot.recipient_color
 
         await destination.trigger_typing()
 
@@ -434,14 +440,19 @@ class ThreadManager:
     async def create(self, recipient, *, creator=None, category=None):
         """Creates a Modmail thread"""
 
-        em = discord.Embed(
-            title='Thread created!',
-            description=self.bot.config.get(
+        thread_creation_response = self.bot.config.get(
                 'thread_creation_response',
                 'The moderation team will get back to you as soon as possible!'
-            ),
-            color=discord.Color.green()
+            )
+
+        em = discord.Embed(
+            color=self.bot.mod_color,
+            description=thread_creation_response,
+            title='Thread Created',
+            timestamp=datetime.utcnow(),
         )
+        em.set_footer(text='Your message has been sent',
+                      icon_url=self.bot.guild.icon_url)
 
         if creator is None:
             self.bot.loop.create_task(recipient.send(embed=em))